strategies/knn/api/requirements.txt
# api
fastapi
uvicorn
python-multipart

# DB + env management
sqlalchemy
python-dotenv

Pillow
requests

# data's libs
pandas 
numpy
matplotlib

# models
faiss-cpu